requirements.txt
fastmcp==2.11.3
sqlite3
pydantic==2.11.7
langchain-voyageai==0.1.6
langchain-community==0.3.27
faiss-cpu==1.11.0.post1
python-dotenv==1.1.1